requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
google-api-python-client==2.111.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openai==1.0.0
Pillow==10.0.1
opencv-python==4.8.1.78
moviepy==1.0.3
